Implement chat functionality with message handling and display components
